laboratorio 14

+++ Lab13C/Models/Detail.cs
@@ -4,8 +4,8 @@
     {
         public int DetailId { get; set; }
         public int Amount { get; set; }
-        public float Price { get; set; }
-        public float SubTotal { get; set; }
+        public decimal Price { get; set; }
+        public decimal SubTotal { get; set; }
 
         public int ProductId { get; set; }
         public Product Product { get; set; }
