+++ Lab13C/Models/Invoice.cs
@@ -5,7 +5,7 @@
         public int InvoiceId { get; set; }
         public DateTime Date { get; set; }
         public string InvoiceNumber { get; set; }
-        public float Total { get; set; }
+        public decimal Total { get; set; }
 
         public int CustomerId { get; set; }
         public Customer Customer { get; set; }
